requirements.txt
bottle==0.12.19
alembic==1.4.3
SQLAlchemy==1.3.20
psycopg2-binary==2.8.6
validate-docbr==1.8.2
